# apps/ade-web/public/openapi/ade-api-openapi.json
{
  "components": {
    "schemas": {
...
      "ApiCommentingGenerateResponse": {
        "properties": {
          "cache_prompt": {
            "title": "Cache Prompt",
            "type": "boolean"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "content_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Source"
          },
          "enable_thinking": {
            "default": false,
            "title": "Enable Thinking",
            "type": "boolean"
          },
          "finish_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finish Reason"
          },
          "max_tokens": {
            "title": "Max Tokens",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "model_key": {
            "title": "Model Key",
            "type": "string"
          },
          "persona_key": {
            "title": "Persona Key",
            "type": "string"
          },
          "prompt_key": {
            "title": "Prompt Key",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "provider_model_id": {
            "title": "Provider Model Id",
            "type": "string"
          },
          "raw_reply": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Reply"
          },
          "raw_request": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Raw Request"
          },
          "received_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Received At"
          },
          "scenario": {
            "enum": [
              "chat",
              "comment",
              "label"
            ],
            "title": "Scenario",
            "type": "string"
          },
          "selected_attempt": {
            "title": "Selected Attempt",
            "type": "string"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_label": {
            "title": "Source Label",
            "type": "string"
          },
          "task_shape": {
            "enum": [
              "classic",
              "all_in_system",
              "structured_output"
            ],
            "title": "Task Shape",
            "type": "string"
          },
          "temperature": {
            "title": "Temperature",
            "type": "number"
          },
          "timeout_seconds": {
            "title": "Timeout Seconds",
            "type": "number"
          },
          "top_k": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Top K"
          },
          "top_p": {
            "title": "Top P",
            "type": "number"
          },
          "usage": {
            "additionalProperties": true,
            "title": "Usage",
            "type": "object"
          }
        },
        "required": [
          "scenario",
          "model_key",
          "source_id",
          "source_label",
          "provider_model_id",
          "prompt_key",
          "persona_key",
          "model",
          "content",
          "provider",
          "max_tokens",
          "timeout_seconds",
          "task_shape",
          "cache_prompt",
          "temperature",
          "top_p",
          "selected_attempt"
        ],
        "title": "ApiCommentingGenerateResponse",
        "type": "object"
      },
...
      "CommentingGenerateRequest": {
        "examples": [
          {
            "cache_prompt": false,
            "enable_thinking": false,
            "input": "Summarize the reader reaction and write one concise editor-style reply.",
            "max_tokens": 512,
            "model_key": "local_llama_server::gemma4",
            "persona_key": "comment_linxiaotang",
            "prompt_key": "comment_v20260418",
            "retry_count": 1,
            "scenario": "comment",
            "task_shape": "classic",
            "temperature": 0.6,
            "timeout_seconds": 120,
            "top_k": 64,
            "top_p": 1.0
          }
        ],
        "properties": {
          "cache_prompt": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "llama.cpp prompt-cache toggle for Comment Lab. Defaults to false and is only sent to llama.cpp-backed sources.",
            "title": "Cache Prompt"
          },
          "enable_thinking": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional vLLM/Gemma thinking toggle. When true for supported vLLM sources, ADE sends `chat_template_kwargs.enable_thinking=true`.",
            "title": "Enable Thinking"
          },
          "include_diagnostics": {
            "default": false,
            "description": "Administrator-only opt-in for raw provider request/reply diagnostics.",
            "title": "Include Diagnostics",
            "type": "boolean"
          },
          "input": {
            "description": "News article, comment thread, or source text to comment on.",
            "title": "Input",
            "type": "string"
          },
          "max_tokens": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional response token budget. Defaults to Comment Lab runtime settings.",
            "title": "Max Tokens"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "deprecated": true,
            "description": "Legacy selector kept for backward compatibility. Prefer `model_key`.",
            "title": "Model"
          },
          "model_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Router-scoped model key from `/api/v2/model-catalog/options?scenario=comment`, for example `local_llama_server::gemma4`.",
            "examples": [
              "local_llama_server::gemma4"
            ],
            "title": "Model Key"
          },
          "persona_key": {
            "default": "comment_linxiaotang",
            "description": "Comment Lab persona key from `/api/v2/model-catalog/options?scenario=comment`.",
            "title": "Persona Key",
            "type": "string"
          },
          "prompt_key": {
            "default": "comment_v20260418",
            "description": "Comment Lab prompt key from `/api/v2/model-catalog/options?scenario=comment`.",
            "title": "Prompt Key",
            "type": "string"
          },
          "retry_count": {
            "anyOf": [
              {
                "maximum": 5.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional provider retry count for transient failures.",
            "title": "Retry Count"
          },
          "scenario": {
            "default": "comment",
            "description": "Must be `comment` for this endpoint.",
            "enum": [
              "chat",
              "comment",
              "label"
            ],
            "title": "Scenario",
            "type": "string"
          },
          "task_shape": {
            "anyOf": [
              {
                "enum": [
                  "classic",
                  "all_in_system",
                  "structured_output"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Prompt-packing strategy. Defaults to the Comment Lab runtime setting.",
            "title": "Task Shape"
          },
          "temperature": {
            "anyOf": [
              {
                "maximum": 2.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Sampling temperature. Defaults to Comment Lab runtime settings.",
            "title": "Temperature"
          },
          "timeout_seconds": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional provider timeout in seconds. Use a realistic local-model value such as 120.",
            "title": "Timeout Seconds"
          },
          "top_k": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional top_k sampling value. Defaults to model profile or Comment Lab runtime settings.",
            "title": "Top K"
          },
          "top_p": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 1.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Nucleus sampling top_p. Defaults to Comment Lab runtime settings.",
            "title": "Top P"
          }
        },
        "required": [
          "input"
        ],
        "title": "CommentingGenerateRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Provides the versioned Agent Studio, lab, content-center, catalog, and test APIs used by ADE Web, workflows, and first-class developer clients.",
    "summary": "Feature-aligned runtime and authoring APIs for Letta Open ADE",
    "title": "ADE API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/comment-lab/generations": {
      "post": {
        "operationId": "api_commenting_generate_api_v2_comment_lab_generations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "llama_server_comment": {
                  "summary": "Generate with local llama-server",
                  "value": {
                    "cache_prompt": false,
                    "enable_thinking": false,
                    "input": "Summarize the reader reaction and write one concise editor-style reply.",
                    "max_tokens": 512,
                    "model_key": "local_llama_server::gemma4",
                    "persona_key": "comment_linxiaotang",
                    "prompt_key": "comment_v20260418",
                    "retry_count": 1,
                    "scenario": "comment",
                    "task_shape": "classic",
                    "temperature": 0.6,
                    "timeout_seconds": 120,
                    "top_k": 64,
                    "top_p": 1.0
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/CommentingGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiCommentingGenerateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Generate a stateless comment for news/comment threads",
        "tags": [
          "Comment Lab"
        ]
      }
    },
    "/api/v2/label-lab/generations": {
      "post": {
        "operationId": "api_labeling_generate_api_v2_label_lab_generations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "examples": {
                "football_entities": {
                  "summary": "Extract football players and teams",
                  "value": {
                    "input": "Messi scored for Inter Miami against Orlando City.",
                    "max_tokens": 1024,
                    "model_key": "local_llama_server::gemma4",
                    "prompt_key": "label_football_entities_v1",
                    "repair_retry_count": 1,
                    "scenario": "label",
                    "schema_key": "label_football_entity_groups_v1",
                    "temperature": 0.0,
                    "timeout_seconds": 120,
                    "top_k": 64,
                    "top_p": 1.0
                  }
                }
              },
              "schema": {
                "$ref": "#/components/schemas/LabelingGenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiLabelingGenerateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Generate stateless grouped entity extraction for an input article",
        "tags": [
          "Label Lab"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local ADE API",
      "url": "http://127.0.0.1:8000"
    }
  ],
...
}